move match to cli
