+++ src/apps.rs
@@ -1,0 +1,1 @@
+pub mod cli;
